backend/requirements.txt
fastapi
uvicorn
gunicorn
pydantic
pydantic-settings
python-multipart
Pillow
numpy
opencv-python-headless
ultralytics